Add delete document; finish delete operation on open dialog.

+++ src/platform/database/index.ts
@@ -1,5 +1,7 @@
-import { UserModel, DocumentModel } from "./models";
+import { UserModel, DocumentModel, ShapeModel } from "./models";
 import { transaction } from "objection";
+import { PubPage } from "../../types/pub-objects";
+import { PageModel } from "./models/page";
 
 const createUser = async function(name, password) {
   const newUser = await UserModel.query().insert({
@@ -31,12 +33,24 @@ const getDocument = function({ userId, id }) {
     .where("id", id);
 };
 
-const deleteDocument = function({ userId, id }) {
-  return DocumentModel.query()
-    .eager("pages.shapes")
-    .delete()
-    .where("user_id", userId)
-    .where("id", id);
+const deleteDocument = async function(options) {
+  return transaction<DocumentModel>(DocumentModel.knex(), async function(trx) {
+    const [document] = await getDocument(options);
+    if (!document) {
+      throw new Error("Document to delete for user does not exist.");
+    }
+    await document.pages.forEach(
+      async (page: PubPage) =>
+        await ShapeModel.query(trx)
+          .delete()
+          .where("page_id", page.id)
+    );
+    await PageModel.query(trx)
+      .delete()
+      .where("document_id", document.id);
+    await DocumentModel.query(trx).deleteById(document.id);
+    return document;
+  });
 };
 
 const getDocuments = function(userId) {
@@ -47,17 +61,10 @@ const getDocuments = function(userId) {
 
 const saveDocument = function(document, userId) {
   return transaction(DocumentModel.knex(), function(trx) {
-    console.log(JSON.stringify(document));
-    return DocumentModel.query(trx).upsertGraphAndFetch(
-      {
-        ...document,
-        userId,
-      }
-      // {
-      //   relate: true,
-      //   unrelate: true,
-      // }
-    );
+    return DocumentModel.query(trx).upsertGraphAndFetch({
+      ...document,
+      userId,
+    });
   });
 };
